refactor: move struct & constructor definition to dedicated file
